+++ src/window/WindowManager.hpp
@@ -5,7 +5,10 @@
 #include <SDL2/SDL.h>
 #include <array>
 #include <cmath>
+#include <cstddef>
 #include <cstdint>
+#include <iterator>
+#include <list>
 #include <SDL2/SDL_ttf.h>
 #include <memory>
 #include <set>